requirements.txt
fastapi
uvicorn[standard]
python-multipart
pydantic
qdrant-client
sentence-transformers
faiss-cpu
pymupdf
pillow
pytesseract
tqdm
rank_bm25
networkx
pyvis
python-dotenv
requests
scikit-learn
spacy
